Initial Commit
This is both pretty rudimentary, and also pretty good!
It reads the script and parses the very limited markdown subset.

Then it wraps it to fit in the given window, and allows me to scroll
around. With italics and bold!

It doesn't yet to any of the things I want it to do, but as a simple
pager it's pretty usable!