Selection
I now have a selection that moves through the file, skipping blank
lines, marking a range of sentences as active.

I can move it up and down, and also extend it. Magic!

Honestly, slightly crazy, but also pretty great.

The thing it doesn't currently do is move the scroll in any way, and it
also doesn't colour the spaces between sentences.