Color Spaces Between Active Sentences
I now represent the inter-sentence space with an object. This allows it
to be activated and deactivated and redrawn. Then I inject it into the
cursor linked-list between the sentences.

If that was all I did, I'd have to move my selection through the spaces
manually, which is annoying.

So I've invented a skippable cursor entry. When going from one to
another, it skips over the skippable entries. But! When iterating, it
doesn't.

This means that going from one sentence to the next works just the same
as it currently does, but when activating and deactivating it still
visits the in-between items.