Basics of Sync
I've built a sync form that allows punching in a time a la the
microwave. It's modal, and then it returns to the main thing.

I'm not _doing_ anything with this yet, but it's there.
It's a step at least.

I tried using the curses form library, but I could not get it to work...
so I just built my own thing.

In the end, the microwave-style input wouldn't have been possible with
the other way anyway, so it all kinda worked out.