DebugDisplay
I'm having a hard time debugging stuff when I can't just puts out some
log lines.

So I've done this now, where I now have a singleton that I can just dump
to from wherever and it'll show up at the bottom of the screen.

But, if I should be able to pull it out later, so it's okay!