Take Management
I can now hit "i" to start a take and "o" to end one.
Multiple "i"s in a row start new takes without ending the old one
(essentially false start)

And then when looking at a given selection it shows which takes included
that selection, even if it was a subset of a previous take's range.

There isn't a lot you can _do_ with the takes, but it's there.
Later you'll be able to mark their quality, but that's about it.

It also doesn't currently do any display of the take's current state,
but I wanted to snapshot here before implementing that.