Take Selection and Marking
Okay great! I can now select between the various takes, and mark the
selected take as poor, okay, good, or trash!

And then that marking is represented with colour.

I also moved the "ITALIC" hack from out of Emph and into the Stylesheet.
Originally this was because I was using it for subset, but that was
actually deemed too subtle, so I went with underline instead. It was a
good move, though, so I left it.

Besides the fact that this whole thing doesn't produce any outputs...
this is _nearly_ feature complete!