Sync Display
I built the sync form before, but didn't use it for anything.
Now I show the current synced time in the top right.
In order for it to count up, I need to set the timeout low enough that
when I'm not issuing any key commands it still ticks.

Also, now that I'm trying to use the result of the sync form, I noticed
that it didn't actually work... So I fixed that too.