Applicationize Application
Previously this was a script.
That's fine!

But I was getting to the point where I wanted a few helper methods in my
"main" flow, but then I'd have to define them above the actual script
logic, and pass state, and the whole thing just got a little wonkier.

So I decided it was complicated enough that it should just be an
explicit object so I can start breaking it up some.

Along the way I noticed that "oopsy, I was reaching into the global
scope and finding doupdate"

Glad this was only in two places, but if nothing else making this scope
change here is going to make it harder to screw this up later by just
importing all of curses into the global scope...