Make Application Instance Variables
This is a step towards what I'm actually working on.
No new features or changes here really, but it touched a bunch of lines,
so I wanted to capture these in one place.

I just replaced a bunch of local variables with Application instance
variables so I can use them later in other methods.