Responsive Layout
This was a bit involved. The gist is that I wanted to be able to handle
resizing after the program started. But doing that in practice was a bit
of work, because all of my layout was so manual.

So! I now initialize all my windows to all be 1x1 and at 0,0 and then
handle the sizing and positioning in the layout method. This is nice
actually because it splits my initialization logic from layout, but it
took a bit of detangling.

It also exposed a few hidden assumptions. Like, that my script flowing
and wrapping was what created the inter-sentence spaces, but the
selection creation is what builds the linked-list.

So if I build the selection before doing layout, I _won't_ have any
spaces in my selection. During layout they'll be put in, but they won't
be selected between sentences anymore because they weren't there during
selection initialization. Which is their whole purpose. A bit tricky to
find!

This is still there, but I'm more careful to build the selection only
after first layout.

Another thing related to these spaces is that when re-flowing the
position of the ends of sentences moves! So I previously created the
spaces conditionally, but now I instead create them all the time, but
only draw them conditionally.

Another weird thing I found was that some of my elements could move as
far as they wanted to the right, but wouldn't work if I tried to move
them left from their starting place...

It turns out that this has something to do with my use of derwin? I
don't know why... but it seems related.
When I made new top-level windows, the problem went away. Maybe derwin
has some kind of bounds-checking on the parent or something that was
corrupting things?

Whatever. I was only using derwin because I liked the idea of it using
relative positions, but I wasn't seeing that benefit in practice since I
was only using it on the root window.
So I just got rid of it.

I was also running derwin inside SyncForm, but none of my later
abstractions were doing that, so I moved some stuff around so I build
the window on the outside and it gets resized on the inside.

All in all, a good thing to have done.