Session Management (Part 1)
I've rebuilt the state-management flow to go through the Session object.
The idea is that all state-changing operations can be streamed out to
the log of all updates. Then it can be replayed and we should be back in
the same state.

To try to make that airtight I've adopted a flow that takes all of the
events, serializes them, and then deserialized them, and then calls the
handlers which update the application state.

This should hopefully make the "live" flow as similar as possible to the
"resume" flow, since it's all the same code. I already found a few bugs
in my deserializer like this, which previously only would have shown up
on the second run of a given session.

The one thing I haven't built yet is the actual stream output and input.
So for now it's local-only, but the live case seems to work!
And, as stated, the code path should be the same.

This does make some of the dataflow a _little_ awkward as it round-trips
a few things through strings and stuff. But again, that's because I want
the live flow to use the resume logic, and so I can't just use in-memory
references to stuff.

The main tweaks were to the TakeManager, because we can't just dump
things into an array in a way that makes sense to resume.
So we have to be much more careful to allocate IDs and verify where our
data is coming from.