Get Selection was Returning Invalid Selections
I was accidentally using the wrong enumerator and was returning a
selection that pointed at two sentences instead of two cursor objects
that pointed to sentences.

This makes the selections unable to traverse, which is kinda their whole
point.

But! It didn't matter! Because the only place we were using
`get_selection` was when building takes. And the only thing we were
_using_ a Take's selection for was testing if a take was inside a range
or not. And testing if something is inside a range was just a comparison
by starting and ending ID. And you *can* get an ID from sentences!

So it transparently worked in the one place I was using it. But when
building something else that used the take's span for something real, it
crashed immediately as soon as I tried to iterate through the selection.