Immutable Selection
Having the selection be immutable has bitten me more than once already,
and then in the next feature I'm building it bit me again.

So no more! They're immutable now, the way God intended.
I honestly didn't have to change too much. I was already treating them
as almost immutable in the way I was handling them, so _mostly_ all I
had to do was make the move and spread commands return something and
that's about it. The way it was being handed around to the WrapPad and
the TakeDisplay was already treating them right.

And while there was a bit of sketchiness with the handling of cursors,
essentially for performance reasons, most of my initial "selection"
building actually got _simpler_ in this new model!
That's a good sign!

And while I was in here anyway, I renamed a few things. I also took out
a dup.freeze, because there was no need! It's always frozen now!

Like I said at the top, I already had several bugs related to this, and
I think I've just nipped a lot of potential future problems in the bud
right here.