Toggle from Subset to Full Take Selection
This allows me to highlight a sentence, then pick a subset take, and
then hit "m" to go to the full selection for that take.

Then I can hit "m" to go right back to where I was. All while keeping
the same take highlighted!

But, if I move the selection around any, it'll reset this toggle and
I'll just be on my own.

Works pretty good, I'd say!