Show Unrecorded Sentences as Dimmer
This makes it much clearer where I am in a script. It doesn't show
anything relating to the quality of the takes or anything, but it does
mean that I can easily see at a glance if I've missed a sentence
entirely, or where I left off, etc.

I was overcomplicating it when I realize that since this "recorded"
state is always "no" until it's "yes", and then never reverts, it made
it much easier. There is no erasing takes, there's no going back. All
forward all the time.