Cleanup Selection Change Handlers
There was starting to be a lot of steps to changing the selection, and
it was maybe a little error prone, but also just a bunch of noise to
scroll through when looking at the handlers.

So I put it all together in one place. I also tweaked it a bit so it can
also handle the toggle concept for me, and manual activation and
deactivation for when I'm jumping around.

Yay!

It does technically mean that it calls `cancel_recording` in a case or
two where I know it can't be recording due to other logic, but that's so
minor it's definitely worth it.

Also, when toggling back to the smaller range, I was explicitly setting
the toggle so I could get back to the main range... but that was
actually not necessary. I don't think I even used it, because in the
case where I was in the subset I just move up like it's the first time.
So I don't set it here either.