Clear Sync
I can now stop the sync timer. Within a session this is usually not
required, but when revisiting a session it would be useless and also
distracting to have a clock that's like: 336:14:32

So now I can stop the clock!
I could maybe make a confirmation dialog... but honestly it's not that
big a deal. I can just reset the sync again and it'll be like nothing
happened. The sync in this system is never the source of truth anyway.

If it becomes a problem in the future, I can build the confirmation at
that time.

Oh! I also changed the keybind from "(r)eset sync" to "(s)ync".
I kept hitting "s" anyway, and "(S)top sync" or whatever made way more
sense than "(R)" for... do the opposite of reset?