Most Recent Take
This is a quick way to jump back to where we left off. Somewhat useful
if I've been scrolling around or something, but much more useful when
resuming a session and wanting to get back into it.

I've made it a toggle because it's easy to just copy from the other
case. Maybe if I do another toggle I'll bake the pattern.

Or maybe switch away from the toggles and just a vim-style "'" binding
to just generally jump "back to where I came from", with maybe a bit of
extra logic to save the take in case it's relevant to the new location.

There is _one_ edge-case here by re-using the same toggle-var.
If I use "." to jump to a range, then I can use "m" to jump back.

Kinda weird feeling... but not harmful I don't think.