Show Selected Take Info
I've used a different pattern from before and made a callback on my
TakeDisplay. I was tired by the idea of going to all the places where I
change a take's selection in various ways and putting some method in all
of these places.

The duplication is making me tired.

So! I just made it so the TakeDisplay fires out the current take, and
then the TakeInfoDisplay just uses that to show the timing.

I felt it looked best when it was lined up with the window, even though
it wasn't in the window. And once I was doing that, it looked weird to
have the make take display have a number dangling "off the edge", so I
moved it to also line up with the window.

And while I was there, I also moved the take display one further so
there's an extra space between the takes and the sync time.

Wowee!