Microsecond Precision
I was a dumb, and didn't think about the fact that obviously Time:to_i
converts things to integers, and integer precision is absolutely not
enough to sync audio. A full second could be the difference between
stopping a recording and starting to talk about how it went.

Definitely it wouldn't be enough to cut out the sound of the keypress
itself!

So yeah, that wasn't going to cut it. The clock was also clearly not in
sync upon coming back.
Even doing this wasn't enough to fix _that_ though, because we also need
to update the display frequency. If the clock only updates every 500ms,
we aren't going to be able to see it ticking at the same rate.

That wouldn't have affected the actual accuracy of the takes, but it
does make it harder to tell how good the sync is.