Sync Scootcher
While I was trying this for real, I noticed that getting the sync right
was tricky. I could get it close, but never bang on.

So I added some commands to slightly shift the sync one way or the other
so I could get it mostly right and then sneak up on it.