Playback of Takes (Part 1)
This was a direction I didn't expect to go in, but the opportunity
presented itself, and it's actually pretty great!
With this I can hit "p" on any take on any line and it'll play me my
take! After recording is done, of course, but that's okay!

It's kind of opening up a whole side I didn't think of before, which is
being able to use this not just during recording, but as a state
_viewer_ after recording. Obviously I dabbled in that a little with
"Clear Sync", which is useful to get the time to go away when you're not
actively recording. But even more than that I found that the easiest way
to reference what I'd recorded was with this tool which had all that
info. Duh!

And with this ability, it's even more useful, not just as a way of
marking takes during recording, but also re-marking takes afterwards and
seeing if they flow well together, etc.

Finding gaps, and maybe even re-recording things if I've got an exported
file playing back while I'm recording a new set of takes. Pretty cool
stuff.

This is part 1, because this isn't a full feature yet. Right now it
involves a hard-coded filename which is what I tested with. This whole
feature was an experiment, and it turned out great, so I'll continue it.
But, putting a filename in involves a form, and validation, and dumping
it into the session, and serializing/deserializing, etc. All important
stuff, but stuff that felt like more work than I wanted to do right now.

Having all of that muddy up the diff, which would then be full of both
file output stuff and file-name form stuff, would just be too confusing
I think. So this is best.

It works with the hard-coded filename, and the rest is a different
feature.