Fine Tune Playback
It turns out even if I match the sync very well, there can still be a
systematic delay between what the counter says and what it records.

So, once I've set up my session for playback, I need to adjust the
output sync or all the clips will cut off words, etc.

So I built this "Fine Tune" widget, and added it to the session.
Then it sets the global sync offset so the playback sounds right!

Currently I just have one global offset, which works so long as the
entirety of my recording is equally well synced, but then there was some
offset anyway. If there are different levels of sync between different
parts of the recording, I'm screwed right now.
I can build that later if I need to, but I'm not sure how I'd even do
that quite yet, so hopefully I just never have to...