Refactor Session Management
There was a lot of boilerplate here, and it was making me not look
forward to adding new events.

Also, while I haven't committed any bugs related to this, because it's a
lot of copy-paste I have made mistakes in development where I forgot to
change two letters here, and now this event looks like that event, etc.

No longer! I designed the abstraction I wanted, and asked ChatGPT to
implement it for me, and it did pretty good! I had to tweak a _few_
things, but this was a lot of boilerplate I could get rid of now.

And as a result, it's also less prone to errors, because it's more
automated.

It also technically has more features, because I can now register
multiple handlers per event. I could have built the original like that,
but I just didn't. Whereas the version ChatGPT built just did that, and
I figured whatever and kept it in!

So no new outside features here, but hopefully this will make it easier
to implement new crap from here on.