Set Playback Filename
Rather than hard-coded, I can now set the name of the recording I play
my lines out of.

There's a bunch of stuff here, but relatively little of it is directly
related to the task, and is more related to general abilities.

Like, for example, I built an entire inputbox implementation that can
take arbitrary strings and scrolls left and right off the edge, etc.
Then I use that in the form, which is nothing more than just a label and
that input box, basically.

So it's related in that it didn't exist and I had to write it, but going
forwards now I expect I'll use this again without having to build it
another time.

I also needed the arrow keys for the first time in the project, which
meant I had to realize that "keypad" is per-window for some reason, and
also works by interpretting the escape characters.
And the default escape time is 1000ms! Which is very noticeable anytime
I tried to close something.

So I had to dial that down to 10, and use keypad right, so that I could
get the arrow keys.
In doing so, I've adjusted my test app while testing.

Other than that I tweaked the safety checks in output to not only have
audio_filename be set to something, but also to a readable file.
I also have to dup the input to inputbox, because `slice` mutates, and I
was editing the value, but cancelling, but it was still messing stuff
up.

And I got to use my new abstraction for session! It was a breeze!
Basically 3 lines and the whole thing was done! So good!