Export Best Takes
I have a bunch of export things I want to do... but I'd have to build a
menu or something to get there.
For now I just wanted to try a thing, which was exporting my stuff in
such a format that Audacity could import it.

It worked great! But I had to lay some groundwork first.

I needed to make it so rather than just testing if one selection was
within another, I could test if a single ID was within a selection.
This is a bit hacky, but it works for now! That allows me to find all
takes that feature a given sentence.

Then, from there, I had to build and ordering into my takes, so they
could be compared too. I was a dummy when I built this thing in that I
put TRSH as the highest number, and have an effective empty status of
zero, which is below everything.
This made sense purely as an enum, but is not a good ordering.

So I made up an ordering that I liked, and used that to implement an
ordering between takes. This allows me to find "the best takes" for a
given sentence.

Then I can dump them out as we go, simply making sure not to output any
take more than once so we've got the case handled where the best take
that spans sentences isn't output for each sentence.

I also have been treating the data inside the Duration objects as an
int, but it's actually a float. Printf didn't know that, though, so I
told it. I should honestly maybe fix all the ".to_i"s, but meh.

And that's about it! If I adjust my output by my global offset and dump
in the format that Audacity can import, and it works great! I get all my
takes lined up with my audio. I may adjust the labels later to be the
sentence range rather than the take number, but all in all this is
pretty great.

And like I said at the top, there are probably other formats I could
want, and I may want to export all takes, or all takes better than OKAY,
or whatever. There's some stuff there, but this is the ground-work at
least!