Play Chosen Takes
I'm honestly not 100% sure about this one... but as a demo it's pretty
good, and maybe I even like it.

Basically, now I can pick a selection of sentences and hitting "p" would
very likely do nothing, because there are no takes that span this range.
But if I hit "P", then it will instead play me the best takes of this
selected range. Essentially allowing me to preview what the export would
export.

I think it's useful as a way of gauging how things are going... but I
could be wrong...
Oh well, I can always delete it later!