Full Assembly Export
Another experiment. I've hard-coded the name for now, because I'm not
sure how I want exports to go generally yet.

But this just takes all the sentences, cuts the best version of each
one, and then assembles them into one long audio file. Essentially doing
the rough assembly for me!

This gives me a sense of how the thing is going, but it also could be
used as a starting point for editing, splitting out all the garbage and
just giving me only the bits I was happiest with, in the order they'll
be in the final work.

If I've got two takes of a single sentence, there may still be some
deduplication that needs to happen, but that's actually the normal part
of the process. Picking between two good takes is art, but having to
listen through 4 aborted takes is not.

Pretty happy so far. If anything this is the culmination of the original
intent for this thing!